Added makefile for easier building
