+++ Cargo.toml
@@ -1,6 +1,8 @@
 [workspace]
 
 members = [
+    "application",
     "application-backend",
-    "application-platform"
+    "application-media",
+    "application-platform",
 ]
