--- application-backend/src/main/java/com/github/yoep/popcorn/backend/ApplicationBackend.java
@@ -1,10 +1,0 @@
-package com.github.yoep.popcorn.backend;
-
-import com.sun.jna.Library;
-import com.sun.jna.Native;
-
-public interface ApplicationBackend extends Library {
-    ApplicationBackend INSTANCE = Native.load("application_backend", ApplicationBackend.class);
-
-    PopcornFx new_instance();
-}
