+++ application-media/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "application-backend"
+name = "application-media"
 version = "0.4.0"
 edition = "2021"
 rust-version = "1.62"
@@ -11,6 +11,10 @@ crate-type = ["cdylib"]
 
 [dependencies]
 log = "0.4.16"
+
+[target.'cfg(windows)'.dependencies]
+windows = { version = "0.39.0", features = [
+    "Media",] }
 
 [build-dependencies]
 cbindgen = "0.24.3"
