+++ application-media/src/popcorn/fx/media/media.rs
@@ -1,0 +1,5 @@
+
+/// The media control provides a way to ley the system now which media is currently being played.
+pub trait MediaControl {
+
+}
