+++ application-media/src/popcorn/fx/media/mod.rs
@@ -1,0 +1,3 @@
+pub mod media;
+#[cfg(target_os = "windows")]
+pub mod media_win;
