+++ application-media/src/popcorn/fx/mod.rs
@@ -1,0 +1,1 @@
+pub mod media;
