+++ application/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "application-backend"
+name = "application"
 version = "0.4.0"
 edition = "2021"
 rust-version = "1.62"
@@ -11,6 +11,7 @@ crate-type = ["cdylib"]
 
 [dependencies]
 log = "0.4.16"
+log4rs = "1.1.1"
 
 [build-dependencies]
 cbindgen = "0.24.3"
