+++ application/src/main/java/com/github/yoep/popcorn/PopcornFx.java
@@ -1,4 +1,4 @@
-package com.github.yoep.popcorn.backend;
+package com.github.yoep.popcorn;
 
 import com.sun.jna.PointerType;
